requirements.txt
streamlit
requests
pandas
python-dotenv
plotly
sqlite3-binary
